remove basic_pyspark_crag and emr_pyspark_crag folders
