+++ docs/content/integrations/spark.mdx
@@ -46,7 +46,7 @@ This example assumes that all the outputs within the dagster job will be PySpark
 
 This example writes out DataFrames to the local file system, but can be tweaked to write to cloud object stores like S3 by changing to the `write` and `read` invocations.
 
-```python file=../../basic_pyspark_crag/repo.py startafter=start_repo_marker_0 endbefore=end_repo_marker_0
+```python file=../../basic_pyspark/repo.py startafter=start_repo_marker_0 endbefore=end_repo_marker_0
 import os
 
 from dagster import IOManager, graph, io_manager, op, repository
@@ -100,7 +100,7 @@ make_and_filter_data_job = make_and_filter_data.to_job(
 
 This example demonstrates how to use the <PyObject module='dagster_aws.emr' object='emr_pyspark_step_launcher' /> to have an op run as a Spark step on an EMR cluster. In it, each of the three ops will be executed as a separate EMR step on the same EMR cluster.
 
-```python file=../../emr_pyspark_crag/repo.py startafter=start-snippet endbefore=end-snippet
+```python file=../../emr_pyspark/repo.py startafter=start-snippet endbefore=end-snippet
 from pathlib import Path
 
 from dagster import graph, make_python_type_usable_as_dagster_type, op, repository
