--- examples/basic_pyspark_crag/requirements.txt
@@ -1,2 +1,0 @@
-dagster
-dagit
