--- examples/basic_pyspark_crag/workspace.yaml
@@ -1,2 +1,0 @@
-load_from:
-  - python_file: repo.py
