--- examples/emr_pyspark_crag/README.md
@@ -1,1 +1,0 @@
-View this example in the Dagster docs at https://docs.dagster.io/examples/emr_pyspark.
