--- examples/emr_pyspark_crag/requirements.txt
@@ -1,4 +1,0 @@
-dagster
-dagit
-dagster-aws
-dagster-pyspark
