--- examples/emr_pyspark_crag/tox.ini
@@ -1,24 +1,0 @@
-[tox]
-envlist = py{38,37,36}-{unix,windows},pylint
-skipsdist = True
-
-[testenv]
-pip_version = pip==21.3.1
-passenv = CI_* COVERALLS_REPO_TOKEN BUILDKITE
-deps =
-  -e ../../python_modules/dagster[test]
-  -e ../../python_modules/libraries/dagster-aws[test]
-  -e ../../python_modules/libraries/dagster-spark
-  -e ../../python_modules/libraries/dagster-pyspark
-whitelist_externals =
-  /bin/bash
-  echo
-commands =
-  !windows: /bin/bash -c '! pip list --exclude-editable | grep -e dagster -e dagit'
-  echo -e "--- \033[0;32m:pytest: Running tox tests\033[0m"
-  pytest -vv {posargs}
-
-[testenv:pylint]
-basepython = python
-commands =
-  /bin/bash -c 'cd .. && pylint -j 0 --rcfile=../.pylintrc emr_pyspark_crag/'
